New endpoint for datasets using the kinds patter
