+++ marquez.dev.yml
@@ -1,17 +1,21 @@
 server:
   applicationConnectors:
-  - type: http
-    port: ${MARQUEZ_PORT:-5000}
-    httpCompliance: RFC7230_LEGACY
+    - type: http
+      port: ${MARQUEZ_PORT:-5000}
+      httpCompliance: RFC7230_LEGACY
   adminConnectors:
-  - type: http
-    port: ${MARQUEZ_ADMIN_PORT:-5001}
+    - type: http
+      port: ${MARQUEZ_ADMIN_PORT:-5001}
 
 db:
   driverClass: org.postgresql.Driver
-  url: jdbc:postgresql://${POSTGRES_HOST:-localhost}:${POSTGRES_PORT:-5432}/marquez
-  user: marquez
-  password: marquez
+  url: jdbc:postgresql://prod-marquez-cluster-aurora-postgresql.cluster-ro-copdpdasx0pj.us-east-1.rds.amazonaws.com/marquez
+  user: marquez_svc
+  password: f74f4f1f27c456c8698T78!3f
+  properties:
+    testOnBorrow: true
+    testWhileIdle: true
+    validationQuery: "SELECT 1"
 
 migrateOnStartup: true
