--- test-new-endpoints.py
@@ -1,257 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script for the new Data Lineage Kinds API endpoints
-"""
-
-import requests
-import json
-import time
-from typing import Dict, Any
-
-def test_conversion_endpoint():
-    """Test the new conversion endpoint with real data"""
-    
-    print("🧪 Testing the new /convert/traditional-to-kinds endpoint")
-    print("=" * 60)
-    
-    # Use the real nodeId we discovered earlier
-    real_node_id = "dataset:ai-core.ingestion:archive.dataset-aws-gcp-mock"
-    
-    conversion_url = "http://localhost:5000/api/graphs/v1alpha1/convert/traditional-to-kinds"
-    
-    payload = {
-        "nodeId": real_node_id,
-        "depth": 3,
-        "targetKind": "LineageGraph",
-        "includeMetadata": True
-    }
-    
-    print(f"📡 POST {conversion_url}")
-    print(f"📋 Payload: {json.dumps(payload, indent=2)}")
-    
-    try:
-        response = requests.post(conversion_url, json=payload, timeout=30)
-        
-        print(f"\n📊 Response Status: {response.status_code}")
-        
-        if response.status_code == 200:
-            data = response.json()
-            
-            print("✅ SUCCESS! Conversion endpoint is working!")
-            
-            # Show conversion metadata
-            conversion_meta = data.get('conversionMetadata', {})
-            print(f"\n🔄 Conversion Metadata:")
-            print(f"   Timestamp: {conversion_meta.get('timestamp', 'unknown')}")
-            print(f"   Source Endpoint: {conversion_meta.get('sourceEndpoint', 'unknown')}")
-            print(f"   Nodes Processed: {conversion_meta.get('nodesProcessed', 0)}")
-            
-            # Show traditional lineage summary
-            traditional = data.get('traditional', {})
-            traditional_nodes = len(traditional.get('graph', []))
-            print(f"\n📈 Traditional Lineage: {traditional_nodes} nodes")
-            
-            # Show kinds lineage summary
-            kinds = data.get('kinds', {})
-            kinds_metadata = kinds.get('metadata', {})
-            kinds_spec = kinds.get('spec', {})
-            kinds_nodes = len(kinds_spec.get('nodes', []))
-            
-            print(f"\n🎯 LineageGraph Kind Summary:")
-            print(f"   API Version: {kinds.get('apiVersion', 'unknown')}")
-            print(f"   Kind: {kinds.get('kind', 'unknown')}")
-            print(f"   Name: {kinds_metadata.get('name', 'unknown')}")
-            print(f"   Graph Depth: {kinds_metadata.get('graphDepth', 'unknown')}")
-            print(f"   Nodes: {kinds_nodes}")
-            
-            # Show central node info
-            central_node = kinds_metadata.get('centralNode', {})
-            if central_node:
-                governance = central_node.get('dataGovernance', {})
-                print(f"\n🎯 Central Node:")
-                print(f"   Name: {central_node.get('name', 'unknown')}")
-                print(f"   Type: {central_node.get('type', 'unknown')}")
-                print(f"   NuRN: {central_node.get('nurn', 'unknown')}")
-                print(f"   Data Domain: {governance.get('dataDomain', 'unknown')}")
-                print(f"   Geo: {governance.get('geo', 'unknown')}")
-            
-            # Show labels and annotations
-            labels = kinds_metadata.get('labels', {})
-            annotations = kinds_metadata.get('annotations', {})
-            
-            if labels:
-                print(f"\n🏷️  Labels:")
-                for key, value in labels.items():
-                    print(f"   {key}: {value}")
-            
-            if annotations:
-                print(f"\n📝 Annotations:")
-                for key, value in annotations.items():
-                    print(f"   {key}: {value}")
-            
-            # Save the full response for inspection
-            output_file = "kinds-conversion-result.json"
-            with open(output_file, 'w') as f:
-                json.dump(data, f, indent=2)
-            print(f"\n💾 Full response saved to: {output_file}")
-            
-            return True
-            
-        else:
-            print(f"❌ Request failed: {response.status_code}")
-            if response.text:
-                print(f"   Response: {response.text}")
-            return False
-            
-    except requests.exceptions.RequestException as e:
-        print(f"❌ Request error: {e}")
-        return False
-    except Exception as e:
-        print(f"❌ Unexpected error: {e}")
-        return False
-
-def test_lineage_graph_endpoint():
-    """Test the lineage-graphs endpoint"""
-    
-    print("\n\n🧪 Testing the /lineage-graphs endpoint")
-    print("=" * 60)
-    
-    # Try to get a lineage graph by name (derived from nodeId)
-    lineage_name = "dataset-ai-core.ingestion-archive.dataset-aws-gcp-mock-lineage"
-    url = f"http://localhost:5000/api/graphs/v1alpha1/lineage-graphs/{lineage_name}"
-    
-    print(f"📡 GET {url}")
-    
-    try:
-        response = requests.get(url, params={"depth": 3}, timeout=30)
-        
-        print(f"📊 Response Status: {response.status_code}")
-        
-        if response.status_code == 200:
-            data = response.json()
-            print("✅ SUCCESS! LineageGraph endpoint is working!")
-            
-            # Show basic info
-            metadata = data.get('metadata', {})
-            spec = data.get('spec', {})
-            
-            print(f"   Name: {metadata.get('name', 'unknown')}")
-            print(f"   Nodes: {len(spec.get('nodes', []))}")
-            
-            return True
-            
-        elif response.status_code == 404:
-            print("ℹ️  404 response (expected - name mapping not implemented yet)")
-            return True  # This is expected for now
-            
-        else:
-            print(f"❌ Unexpected status: {response.status_code}")
-            return False
-            
-    except requests.exceptions.RequestException as e:
-        print(f"❌ Request error: {e}")
-        return False
-
-def test_list_endpoint():
-    """Test the list lineage-graphs endpoint"""
-    
-    print("\n\n🧪 Testing the /lineage-graphs list endpoint")
-    print("=" * 60)
-    
-    url = "http://localhost:5000/api/graphs/v1alpha1/lineage-graphs"
-    
-    print(f"📡 GET {url}")
-    
-    try:
-        response = requests.get(url, params={"limit": 10}, timeout=30)
-        
-        print(f"📊 Response Status: {response.status_code}")
-        
-        if response.status_code == 200:
-            data = response.json()
-            print("✅ SUCCESS! List endpoint is working!")
-            
-            # Show basic info
-            metadata = data.get('metadata', {})
-            items = data.get('items', [])
-            
-            print(f"   API Version: {data.get('apiVersion', 'unknown')}")
-            print(f"   Kind: {data.get('kind', 'unknown')}")
-            print(f"   Total Count: {metadata.get('totalCount', 0)}")
-            print(f"   Items Returned: {len(items)}")
-            
-            return True
-            
-        else:
-            print(f"❌ Request failed: {response.status_code}")
-            return False
-            
-    except requests.exceptions.RequestException as e:
-        print(f"❌ Request error: {e}")
-        return False
-
-def check_server_startup():
-    """Check if the server restarted properly with our new endpoints"""
-    
-    print("🔄 Checking if server needs restart...")
-    
-    # Check basic health
-    try:
-        response = requests.get("http://localhost:5000/api/v1/namespaces", timeout=5)
-        if response.status_code == 200:
-            print("✅ Marquez server is running")
-            return True
-    except:
-        pass
-    
-    print("❌ Server might need restart")
-    print("\n💡 To restart with new endpoints:")
-    print("   ./gradlew run")
-    print("   # or")
-    print("   docker-compose restart")
-    
-    return False
-
-def run_all_tests():
-    """Run all endpoint tests"""
-    
-    print("🚀 Testing New Data Lineage Kinds API Endpoints")
-    print("=" * 70)
-    
-    if not check_server_startup():
-        return
-    
-    # Wait a moment for server to be ready
-    time.sleep(2)
-    
-    success_count = 0
-    total_tests = 3
-    
-    # Test conversion endpoint (most important)
-    if test_conversion_endpoint():
-        success_count += 1
-    
-    # Test individual lineage graph endpoint  
-    if test_lineage_graph_endpoint():
-        success_count += 1
-    
-    # Test list endpoint
-    if test_list_endpoint():
-        success_count += 1
-    
-    print(f"\n" + "=" * 70)
-    print(f"🎉 Test Results: {success_count}/{total_tests} endpoints working")
-    
-    if success_count >= 1:
-        print("\n✅ SUCCESS! Your kinds API is working!")
-        print("\n🎯 What you can do now:")
-        print("1. ✅ Convert traditional lineage to kinds format")
-        print("2. 🧪 Test with different nodeIds and depths")
-        print("3. 🔧 Implement better governance metadata extraction")
-        print("4. 📊 Add the endpoints to your OpenAPI spec")
-        print("5. 🎨 Build UI components that consume these kinds")
-    else:
-        print("\n❌ Issues detected. Check server logs for details.")
-
-if __name__ == "__main__":
-    run_all_tests()
